refactor(studio): split plugin context into context/provider/utils/consts
Move the pure manifest logic (fetchPlugins, loadPlugin, validators) into
utils.ts and the constants (endpoint, query key, empty defaults) into consts.ts.
Separate the PluginProvider component from the context object and hooks
(PluginContext is now a .ts module) so hooks no longer share a file with a
component — dropping the per-export react-refresh/only-export-components disables.
