+++ web/packages/studio/src/routes/index.tsx
@@ -5,7 +5,7 @@ import { ErrorMessage } from '@nemo/common/src/components/ErrorMessage';
 import { ErrorPanel } from '@studio/components/ErrorPanel';
 import { Loading } from '@studio/components/Layouts/Loading';
 import { ROUTES } from '@studio/constants/routes';
-import { PluginProvider } from '@studio/plugins/PluginContext';
+import { PluginProvider } from '@studio/plugins/PluginProvider';
 import { PluginRenderer } from '@studio/plugins/PluginRenderer';
 import {
   agentRoutes,
